🐛Fix: Fix Jwt validateToken return type
