+++ src/main/kotlin/team5/backoffice/infra/security/jwt/JwtPlugin.kt
@@ -1,0 +1,50 @@
+package team5.backoffice.infra.security.jwt
+
+import io.jsonwebtoken.Claims
+import io.jsonwebtoken.Jws
+import io.jsonwebtoken.Jwts
+import io.jsonwebtoken.security.Keys
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.stereotype.Component
+import java.nio.charset.StandardCharsets
+import java.time.Duration
+import java.time.Instant
+import java.util.*
+
+@Component
+class JwtPlugin(
+    @Value("\${auth.jwt.issuer}") private val issuer: String,
+    @Value("\${auth.jwt.secret}") private val secret: String,
+    @Value("\${auth.jwt.accessTokenExpirationHour}") private val accessTokenExpirationHour: Long,
+) {
+    fun validateToken(jwt: String): Result<Jws<Claims>> {
+        return kotlin.runCatching {
+            val key = Keys.hmacShaKeyFor(secret.toByteArray(StandardCharsets.UTF_8))
+            Jwts.parser().verifyWith(key).build().parseSignedClaims(jwt)
+        }
+    }
+
+    fun generateAccessToken(subject: String, email: String, role: String): String {
+        return generateToken(subject, role, email, Duration.ofHours(accessTokenExpirationHour))
+    }
+
+    private fun generateToken(subject: String, role: String, email: String, expirationPeriod: Duration?): String {
+        val claims: Claims = Jwts.claims()
+            .add(mapOf("role" to role))
+            .add(mapOf("email" to email))
+            .build()
+
+        val key = Keys.hmacShaKeyFor(secret.toByteArray(StandardCharsets.UTF_8))
+        val now = Instant.now()
+
+        return Jwts.builder()
+            .subject(subject)
+            .issuer(issuer)
+            .issuedAt(Date.from(now))
+            .expiration(Date.from(now.plus(expirationPeriod)))
+            .claims(claims)
+            .signWith(key)
+            .compact()
+    }
+
+}
